+++ Fans/FanControl.cs
@@ -21,6 +21,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.ComponentModel.Composition;
+using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -58,12 +59,12 @@ namespace DaleGhent.NINA.PlaneWaveTools.Fans {
             CopyMetaData(copyMe);
         }
 
-        public override async Task Execute(IProgress<ApplicationStatus> progress, CancellationToken token) {
+        public override async Task Execute(IProgress<ApplicationStatus> progress, CancellationToken ct) {
             string url = $"{pwi3UrlBase}&clientId={Pwi3ClientId}";
             url += $"&mode={(FanState ? "normal" : "off")}";
 
             try {
-                await Utilities.HttpGetRequestAsync(Pwi3IpAddress, Pwi3Port, url, token);
+                await Utilities.HttpRequestAsync(Pwi3IpAddress, Pwi3Port, url, HttpMethod.Get, string.Empty, ct);
             } catch {
                 throw;
             }
@@ -73,10 +74,6 @@ namespace DaleGhent.NINA.PlaneWaveTools.Fans {
 
         public override object Clone() {
             return new FanControl(this) {
-                Icon = Icon,
-                Name = Name,
-                Category = Category,
-                Description = Description,
                 FanState = FanState,
             };
         }