+++ Properties/AssemblyInfo.cs
@@ -14,7 +14,6 @@ using System.Runtime.InteropServices;
 // [MANDATORY] A short description of your plugin
 [assembly: AssemblyDescription("A collection of useful tools for managing PlaneWave telescope systems")]
 
-
 // The following attributes are not required for the plugin per se, but are required by the official manifest meta data
 
 // Your name
@@ -24,7 +23,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCopyright("Copyright © 2021 Dale Ghent")]
 
 // The minimum Version of N.I.N.A. that this plugin is compatible with
-[assembly: AssemblyMetadata("MinimumApplicationVersion", "2.0.0.2001")]
+[assembly: AssemblyMetadata("MinimumApplicationVersion", "2.0.0.2031")]
 
 // The license your plugin code is using
 [assembly: AssemblyMetadata("License", "MPL-2.0")]
@@ -32,7 +31,6 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyMetadata("LicenseURL", "https://www.mozilla.org/en-US/MPL/2.0/")]
 // The repository where your pluggin is hosted
 [assembly: AssemblyMetadata("Repository", "https://github.com/daleghent/nina-planewave-tools")]
-
 
 // The following attributes are optional for the official manifest meta data
 
@@ -52,16 +50,19 @@ using System.Runtime.InteropServices;
 //[Optional] An additional url to an example example screenshot of your plugin in action
 [assembly: AssemblyMetadata("AltScreenshotURL", "")]
 //[Optional] An in-depth description of your plugin
-[assembly: AssemblyMetadata("LongDescription", @"PlaneWave Tools provides several instructions for controlling the EFA fans and Delta T heater controller via PlaneWave's PWI3 app.
+[assembly: AssemblyMetadata("LongDescription", @"PlaneWave Tools provides several instructions for controlling various aspects of PlaneWave OTA and mount systems.
 
 # Requirements #
 
 * PWI3 >= 3.5.0 for fan control
 * PWI3 >= 3.5.3 for fan and DeltaT control
+* PWI4 >= 4.0.11 beta 9
 
 # Sequence Instructions #
 
 * Start PWI3, Stop PWI3 - Starts and stops the PWI3.exe application
+* Start PWI4, Stop PWI4 - Starts and stops the PWI4.exe application
+* Axis Control - Enable or disable mount axes
 * Fan Control - Turns the OTA fans on or off
 * DeltaT Control - Sets the per-heater operating modes of the Delta T heater controller