+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ﻿# PlaneWave Tools
 
-## 1.1.0.0 - 2021-12-??
+## 1.1.0.0 - 2022-01-09
+* Added Start/Stop PWI4 instructions
+* Added **Set Axis State** instruction for enabling and disabling mount axes via PWI4
 * Changed formatting of DeltaT combobox contents
-* Minimum supported NINA version is now 2.0 Beta 11
+* Updated icons
+* Minimum supported NINA version is now 2.0 Beta 31
 
 ## 1.0.5.0 - 2021-11-14
 * Minimum supported NINA version is now 2.0 Beta 1